requirements.txt
Whoosh==2.7.4
Django==1.11.20
PyMySQL==0.9.3
selenium==3.141.0
redis==3.2.1
requests==2.21.0
jieba==0.39
dwebsocket==0.5.10
pycrypto==2.6.1
Pillow==6.0.0
django_haystack==2.8.1
djangorestframework==3.9.2
